macOS dock icon (needs testing)
Adds code snippet in VenusUI constructor to set the app's dock icon on macOS. To successfully execute, the user must add a command-line argument to the java executable, otherwise an IllegalAccessError is thrown (and caught). The jar should theoretically run on Windows without issue, but I'm not sure if it will build on a Windows machine. Worst case, it uses Java Reflection to load the class dynamically instead of hard-coding it.