Fix up tools, update maven plugins, extract svg action item loading
General tool cleanup. Removes Mars Bot as the UI is broken and is a nonfunctional and old demo tool.
Updates maven plugins to the latest to remove log spam when packaging.
Extracts SVG action item retrieval into SVGIcon as a static method and uses it in ScavengerHunt to fix UI crashing - pending on whether ScavengerHunt should also be removed.