Add syscall 60 (Print I/O Message)
May have to be changed to instead just return the string value at an address? Problem is that buffer is not of fixed size, so may have to be allocated in heap. It is also harder to work with than just syscalling once. Perhaps it can just act like `exception.printStackTrace();`? (Maybe it can take a file descriptor to write to, for more versatile logging)